[vme-test] Provide time in VME readout
Tag each readout with a timestamp. Although timestamps are provided in
milliseconds, actual precision is lower since the data is tagged only
after every board of the same kind has been readout.

This patch is indended for standalone tests only and should not be
commited as is to the main DAQ.